requirements.txt
streamlit
torch
transformers
sentence-transformers
chromadb
pyyaml
pydantic
numpy
llama-cpp-python
accelerate
bitsandbytes